Move get properties from resource to utils
